refactor: move `fetchCoursesByIds` to queries
